refactor raw file
