+++ composables/parser/template/javascript.js
@@ -1,0 +1,8 @@
+function fibonacci(num, memo) {
+  memo = memo || {}
+
+  if (memo[num]) return memo[num]
+  if (num <= 1) return 1
+
+  return (memo[num] = fibonacci(num - 1, memo) + fibonacci(num - 2, memo))
+}
