--- composables/parser/template/javascript.ts
@@ -1,8 +1,0 @@
-export const jsTemplate = `function fibonacci(num, memo) {
-  memo = memo || {};
-
-  if (memo[num]) return memo[num];
-  if (num <= 1) return 1;
-
-  return memo[num] = fibonacci(num - 1, memo) + fibonacci(num - 2, memo);
-}`
